fix(uxrce_dds_client): include counter context in disconnect log lines
Issue #24770 and many duplicates report users seeing "Running, disconnected"
or generic "disconnecting" messages with no diagnostic data — operator has
no way to tell which condition fired.

Each of the three disconnect paths now logs the counter value that crossed
the threshold, the configured threshold itself, and elapsed-time / last-rate
context. Post-mortem from dmesg or SD log is now unambiguous:

  No ping response (missed=3, last_reply_age_ms=3047), disconnecting
  Payload TX rate zero for 5s (limit 3s, last_rate=0 B/s), disconnecting
  Payload RX rate zero for 12s (limit 10s, last_rate=0 B/s), disconnecting

Branch: uxrce-dds-stability-2026-04-30
References:
  https://github.com/PX4/PX4-Autopilot/issues/24770

+++ src/lib/timesync/Timesync.cpp
@@ -110,10 +110,10 @@ void Timesync::update(const uint64_t now_us, const int64_t remote_timestamp_ns, 
 			// high-RTT sample (visible by default), then again every 50 thereafter, and use
 			// PX4_WARN so it shows up in the dmesg ring without enabling debug builds.
 			if (_high_rtt_count == 1 || _high_rtt_count % 50 == 0) {
-				PX4_WARN("timesync: RTT %llu ms exceeds gate %llu ms (count=%u)",
+				PX4_WARN("timesync: RTT %llu ms exceeds gate %llu ms (count=%lu)",
 					 (unsigned long long)(rtt_us / 1000ULL),
 					 (unsigned long long)(MAX_RTT_SAMPLE / 1000ULL),
-					 _high_rtt_count);
+					 (unsigned long)_high_rtt_count);
 			}
 		}
